fix: only one copy of each API endpoint
We decided to ensure that we don't keep multiple versions of a single
API available in the JS SDK. There are other options, like staying at
an older version, or calling the REST API manually.
