+++ src/public-api/model.ts
@@ -18,7 +18,7 @@ export { Webhook } from '../lib/model/webhook.js';
 export { SimpleError } from '../lib/model/errors/simple-error.js';
 export { ComplexError } from '../lib/model/errors/complex-error.js';
 export { UnknownError } from '../lib/model/errors/unknown-error.js';
-export { TicketCreationRequestV2 } from '../lib/model/ticket/ticket-creation-request-v2.js';
+export { TicketCreationRequest } from '../lib/model/ticket/ticket-creation-request.js';
 export { TicketCreatedResponse } from '../lib/model/ticket/ticket-created-response.js';
 export {
   InputFieldRequest,
